+++ iphone/Maps/Common/MWMTypes.h
@@ -15,9 +15,9 @@ typedef NS_ENUM(NSUInteger, MWMTheme) {
   MWMThemeAuto
 };
 
-typedef uint32_t MWMMarkID;
-typedef uint32_t MWMLineID;
-typedef uint32_t MWMMarkGroupID;
+typedef uint64_t MWMMarkID;
+typedef uint64_t MWMLineID;
+typedef uint64_t MWMMarkGroupID;
 typedef NSArray<NSNumber *> * MWMGroupIDCollection;
 
 typedef NS_ENUM(NSUInteger, MWMBookmarksShareStatus) {
